--- Modules/Core/Network/Placeholder2.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  CompanionNote
-//
-//  Created by 김영훈 on 3/19/25.
-//
-
-import Foundation
